NAV-25017: Fjerner behandle klage toggle
