+++ src/main/kotlin/no/nav/familie/ba/sak/ekstern/restDomene/RestJournalføring.kt
@@ -39,10 +39,8 @@ data class RestJournalføring(
     val journalpostTittel: String?,
     val kategori: BehandlingKategori?,
     val underkategori: BehandlingUnderkategori?,
-    val knyttTilFagsak: Boolean?, // TODO : Fjern denne når togglen "familie-ba-sak.klage" er borte
     val opprettOgKnyttTilNyBehandling: Boolean,
     val tilknyttedeBehandlinger: List<TilknyttetBehandling> = emptyList(),
-    val tilknyttedeBehandlingIder: List<String>, // TODO : Fjern denne når togglen "familie-ba-sak.klage" er borte
     val dokumenter: List<RestJournalpostDokument>,
     // Saksbehandler sin ident
     val navIdent: String,
@@ -50,7 +48,6 @@ data class RestJournalføring(
     val nyBehandlingsårsak: BehandlingÅrsak,
     val fagsakType: FagsakType,
     val institusjon: RestInstitusjon? = null,
-    val fagsakId: Long? = null,
 ) {
     fun oppdaterMedDokumentOgSak(
         sak: Sak,
