+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/integrasjoner/journalføring/JournalføringControllerTest.kt
@@ -2,35 +2,23 @@ package no.nav.familie.ba.sak.integrasjoner.journalføring
 
 import io.mockk.every
 import io.mockk.mockk
-import no.nav.familie.ba.sak.config.FeatureToggle
-import no.nav.familie.ba.sak.config.featureToggle.UnleashNextMedContextService
 import no.nav.familie.ba.sak.datagenerator.lagTilgangsstyrtJournalpost
 import no.nav.familie.ba.sak.sikkerhet.TilgangService
 import no.nav.familie.kontrakter.felles.PersonIdent
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.springframework.http.HttpStatus
 
 class JournalføringControllerTest {
-    private val innkommendeJournalføringService: InnkommendeJournalføringService = mockk()
-    private val innkommendeJournalføringServiceV2 = mockk<InnkommendeJournalføringServiceV2>()
+    private val innkommendeJournalføringService = mockk<InnkommendeJournalføringService>()
     private val tilgangService: TilgangService = mockk()
-    private val unleashService = mockk<UnleashNextMedContextService>()
 
     private val journalføringController: JournalføringController =
         JournalføringController(
             innkommendeJournalføringService = innkommendeJournalføringService,
-            innkommendeJournalføringServiceV2 = innkommendeJournalføringServiceV2,
             tilgangService = tilgangService,
-            unleashService = unleashService,
         )
-
-    @BeforeEach
-    fun init() {
-        every { unleashService.isEnabled(FeatureToggle.BEHANDLE_KLAGE, false) } returns true
-    }
 
     @Nested
     inner class HentJournalposterForBruker {
